Move lambda tests into a folder
