+++ serverless/test/lambda/redirect.spec.ts
@@ -4,8 +4,8 @@ import {
   JS_HANDLER,
   PYTHON_HANDLER,
   DD_HANDLER_ENV_VAR,
-} from "../src/lambda/redirect";
-import { LambdaFunction, RuntimeType } from "../src/lambda/layer";
+} from "../../src/lambda/redirect";
+import { LambdaFunction, RuntimeType } from "../../src/lambda/layer";
 
 function mockLambdaFunction(key: string, runtime: string, runtimeType: RuntimeType) {
   return {
